moved effect applying code to new file
