+++ strapi/src/index.ts
@@ -1,6 +1,8 @@
 import { Server, Namespace, Socket } from 'socket.io';
 import { PendingAction, Action, TeamRole, User, PendingActionRequest, ActionType, ActionCompleteRequest, Message } from './types';
 import { DefaultEventsMap } from 'socket.io/dist/typed-events';
+import { getUser } from './utilities';
+import applyEffects from './effects';
 type SocketServer = Server<DefaultEventsMap, DefaultEventsMap, DefaultEventsMap, any> | Namespace<DefaultEventsMap, DefaultEventsMap, DefaultEventsMap, any>;
 
 const FRONTEND_URL = process.env.FRONTEND_URL || 'http://localhost:3000';
@@ -63,31 +65,6 @@ async function checkReceiver(userId: number, receiver: string) {
     }
   });
   return teammates.length > 0;
-}
-
-/**
- * Gets a user's info from their username
- * @param username The user's username
- * @returns The user's info if they exist, otherwise `null`
- */
-async function getUser(username: string) {
-  const res = await strapi.entityService.findMany('plugin::users-permissions.user', {
-    filters: {
-      username: username
-    },
-    populate: '*'
-  });
-  if (res.length === 0) {
-    console.error('user ' + username + ' does not exist');
-    return null;
-  }
-  const user: User = {
-    username: res[0].username,
-    email: res[0].email,
-    teamRole: res[0].teamRole as TeamRole,
-    team: res[0].team.name
-  };
-  return user;
 }
 
 /**
@@ -165,147 +142,8 @@ async function actionComplete(actionCompleteRequest: ActionCompleteRequest, fron
   await strapi.entityService.delete('api::pending-action.pending-action', actionCompleteRequest.pendingActionId);
 
   // parse and apply action effects
-  const effects = (await strapi.entityService.findOne('api::action.action', pendingActionRes.actionId, {
-    populate: ['effects']
-  })).effects;
-
   const user = await getUser(pendingActionRes.user);
-
-  const playerTeam = (await strapi.entityService.findMany('api::team.team', {
-    filters: {
-      name: user.team
-    },
-    populate: '*'
-  }))[0];
-
-  const otherTeam = (await strapi.entityService.findMany('api::team.team', {
-    filters: {
-      $not: {
-        name: user.team
-      }
-    },
-    populate: '*'
-  }))[0];
-
-  effects.forEach(async (effect) => {
-    switch (effect.__component) {
-
-      // add victory points to user's team or opposing team
-      case 'effects.add-victory-points':
-        console.log('EFFECT: adding victory points');
-        if (effect.myTeam) {
-          // add victory points to user's team
-          await strapi.entityService.update('api::team.team', playerTeam.id, {
-            data: {
-              victoryPoints: playerTeam.victoryPoints + effect.points
-            }
-          });
-        } else {
-          // add victory points to opposing team
-          await strapi.entityService.update('api::team.team', otherTeam.id, {
-            data: {
-              victoryPoints: otherTeam.victoryPoints + effect.points
-            }
-          });
-        }
-        break;
-
-      // add a buff or debuff to user
-      case 'effects.buff-debuff':
-        console.log('EFFECT: buffing/debuffing');
-        const id = effect.myTeam ? playerTeam.id : otherTeam.id;
-
-        // probably a better way to do this
-        switch (effect.teamRole) {
-          case 'leader':
-            await strapi.entityService.update('api::team.team', id, {
-              data: {
-                leaderModifiers: {
-                  offense: playerTeam.leaderModifiers.offense,
-                  defense: playerTeam.leaderModifiers.defense,
-                  buff: playerTeam.leaderModifiers.buff + effect.buff
-                }
-              }
-            });
-            break;
-          case 'intelligence':
-            await strapi.entityService.update('api::team.team', id, {
-              data: {
-                intelligenceModifiers: {
-                  offense: playerTeam.intelligenceModifiers.offense,
-                  defense: playerTeam.intelligenceModifiers.defense,
-                  buff: playerTeam.intelligenceModifiers.buff + effect.buff
-                }
-              }
-            });
-            break;
-          case 'military':
-            await strapi.entityService.update('api::team.team', id, {
-              data: {
-                militaryModifiers: {
-                  offense: playerTeam.militaryModifiers.offense,
-                  defense: playerTeam.militaryModifiers.defense,
-                  buff: playerTeam.militaryModifiers.buff + effect.buff
-                }
-              }
-            });
-            break;
-          case 'diplomat':
-            await strapi.entityService.update('api::team.team', id, {
-              data: {
-                diplomatModifiers: {
-                  offense: playerTeam.diplomatModifiers.offense,
-                  defense: playerTeam.diplomatModifiers.defense,
-                  buff: playerTeam.diplomatModifiers.buff + effect.buff
-                }
-              }
-            });
-            break;
-          case 'media':
-            await strapi.entityService.update('api::team.team', id, {
-              data: {
-                mediaModifiers: {
-                  offense: playerTeam.mediaModifiers.offense,
-                  defense: playerTeam.mediaModifiers.defense,
-                  buff: playerTeam.mediaModifiers.buff + effect.buff
-                }
-              }
-            });
-            break;
-        }
-        break;
-
-      // stop an offense action
-      case 'effects.stop-offense-action':
-        console.log('EFFECT: stopping offense action');
-        const res = await strapi.entityService.findMany('api::pending-action.pending-action', {
-          filters: {
-            action: {
-              type: 'offense',
-              teamRole: effect.teamRole
-            }
-          },
-          populate: '*'
-        });
-        const offenseAction = res.filter(async (action) => {
-          const actionUser = await getUser(action.user);
-          return actionUser.team !== user.team;
-        })[0];
-        if (offenseAction) {
-          await strapi.entityService.create('api::resolved-action.resolved-action', {
-            data: {
-              user: offenseAction.user,
-              date: new Date(),
-              action: offenseAction.action,
-              endState: 'stopped'
-            }
-          });
-          await strapi.entityService.delete('api::pending-action.pending-action', offenseAction.id);
-          gameLogic.emit('deleteAction', offenseAction.id);
-        }
-        break;
-    }
-  });
+  await applyEffects(pendingActionRes.actionId, user, gameLogic);
 
   // unlock queue
   gameLogic.emit('queueUnlock');
